requirements.txt
discord.py==2.3.2
fbchat==1.9.7
aiohttp==3.9.5
Flask==3.0.3